+++ packages/app/components/Bouncers/Data.ts
@@ -1,0 +1,22 @@
+import db from "@cuppazee/types";
+
+export type BouncerOverviewData = {
+  [key: string]: number;
+}
+
+export default function BouncerOverviewConverter(data: BouncerOverviewData) {
+  const uncategoriesTypes = new Set<string>();
+  const output: {[key: string]: number} = {};
+  for (const entry of Object.entries(data)) {
+    if (!db.getType(entry[0])) {
+      uncategoriesTypes.add(entry[0]); 
+      output[entry[0]] = entry[1];
+    } else {
+      output[db.strip(entry[0])] = entry[1];
+    }
+  }
+  return {
+    counts: output,
+    uncategoriesTypes: Array.from(uncategoriesTypes),
+  };
+}