+++ packages/app/screens/Tools/Bouncers.tsx
@@ -2,12 +2,15 @@ import db, { TypeHidden, TypeTags } from "@cuppazee/types";
 import { Layout, Text } from "@ui-kitten/components";
 import * as React from "react";
 import { Image, ScrollView, View } from "react-native";
+import BouncerOverviewConverter from "../../components/Bouncers/Data";
+import { BouncerIcon } from "../../components/Bouncers/Icon";
 import useCuppaZeeRequest from "../../hooks/useCuppaZeeRequest";
 import useTitle from "../../hooks/useTitle";
 
 export default function BouncersScreen() {
   useTitle(`☕ Bouncers`);
   const data = useCuppaZeeRequest("bouncers/overview", {});
+  const d = React.useMemo(() => data.data ? BouncerOverviewConverter(data.data.data) : null, [data.dataUpdatedAt]);
   return (
     <Layout style={{ flex: 1 }}>
       <ScrollView
@@ -17,7 +20,22 @@ export default function BouncersScreen() {
           alignItems: "flex-start",
           flexWrap: "wrap",
         }}>
+        {d && d.uncategoriesTypes.length > 0 && (
+          <View style={{ flexGrow: 1, width: 400, maxWidth: "100%", padding: 4 }}>
+            <Layout level="3" style={{ borderRadius: 8, padding: 4 }}>
+              <Text category="h5" style={{ textAlign: "center" }}>
+                Uncategorised
+              </Text>
+              <View style={{ flexDirection: "row", flexWrap: "wrap" }}>
+                {d?.uncategoriesTypes.map(t => (
+                  <BouncerIcon count={d.counts[t]} icon={t} />
+                ))}
+              </View>
+            </Layout>
+          </View>
+        )}
         {db.categories
+          .filter(i => i.active)
           .filter(i =>
             i.types.some(i => i.has_tag(TypeTags.Bouncer) || i.has_tag(TypeTags.Scatter))
           )
@@ -27,13 +45,11 @@ export default function BouncersScreen() {
                 <Text category="h5" style={{ textAlign: "center" }}>
                   {i.name}
                 </Text>
-                <View style={{ flexDirection: "row", flexWrap: "wrap" }}>
+                <View style={{ flexDirection: "row", flexWrap: "wrap", justifyContent: "center" }}>
                   {i.types
                     .filter(i => !i.hidden(TypeHidden.Bouncers))
                     .map(t => (
-                      <View>
-                        <Image style={{ height: 32, width: 32 }} source={{ uri: i.iconURL }} />
-                      </View>
+                      <BouncerIcon count={d?.counts[t.strippedIcon] || 0} type={t} />
                     ))}
                 </View>
               </Layout>